Reorganized caches into their own files
