+++ src/cache/abstract-assignment-cache.ts
@@ -1,8 +1,5 @@
 import { getMD5Hash } from '../obfuscation';
 
-import {ExpiringLRUCache, LRUCache} from './lru-cache';
-import {max} from "lodash";
-
 /**
  * Assignment cache keys are only on the subject and flag level, while the entire value is used
  * for uniqueness checking. This way that if an assigned variation or bandit action changes for a
@@ -79,35 +76,5 @@ export abstract class AbstractAssignmentCache<T extends Map<string, string>>
    */
   entries(): IterableIterator<[string, string]> {
     return this.delegate.entries();
-  }
-}
-
-/**
- * A cache that never expires.
- *
- * The primary use case is for client-side SDKs, where the cache is only used
- * for a single user.
- */
-export class NonExpiringInMemoryAssignmentCache extends AbstractAssignmentCache<
-  Map<string, string>
-> {
-  constructor(store = new Map<string, string>()) {
-    super(store);
-  }
-}
-
-/**
- * A cache that uses the LRU algorithm to evict the least recently used items.
- *
- * It is used to limit the size of the cache.
- *
- * The primary use case is for server-side SDKs, where the cache is shared across
- * multiple users. In this case, the cache size should be set to the maximum number
- * of users that can be active at the same time.
- * @param {number} maxSize - Maximum cache size
- */
-export class LRUInMemoryAssignmentCache extends AbstractAssignmentCache<LRUCache> {
-  constructor(maxSize: number) {
-    super(new LRUCache(maxSize));
   }
 }
